requirements.txt
fastapi
uvicorn
openai
twilio
python-dotenv
requests
pillow
qrcode
aiohttp
python-multipart
httpx
